lib: Allow setting DA address from hardware code information
Mediatek platforms might use different addresses to load/execute the
Download Agent (DA), which can be determined from known hardware codes.

Currently, this DA address is set in the handshake operation, but the
hardware code information is retrieved afterwards.

This commit moves out the DA address from the handshake operation, to be a
parameter of the upload/run DA ones. Also adds a helper function to do the
hwcode to DA address translation.

Signed-off-by: Ariel D'Alessandro <ariel.dalessandro@collabora.com>